+++ src/main/java/com/github/theredbrain/foodoverhaul/block/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.github.theredbrain.foodoverhaul.block;
+
+import org.jspecify.annotations.NullMarked;